requirements.txt
ccxt==4.1.77
pandas==2.1.4
numpy==1.24.3
python-binance==1.0.19
ta==0.10.2
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
websocket-client==1.6.4
